refactor: :recycle: rename type.dart to enum.dart
